refactor: auxiliary to utils folder
