+++ src/general_manager/utils/filterParser.py
