+++ src/general_manager/utils/formatString.py
