+++ src/general_manager/utils/jsonEncoder.py
